+++ sockeye/optimizers.py
@@ -17,7 +17,6 @@ import mxnet as mx
 
 from . import config
 from .lr_scheduler import LearningRateScheduler
-from sockeye_contrib.optimizers import bert_adam
 
 
 class OptimizerConfig(config.Config):