+++ test/unit/test_arguments.py
@@ -16,12 +16,19 @@ import pytest
 import tempfile
 import os
 import re
-import yaml
 
 import sockeye.arguments as arguments
 import sockeye.constants as C
 
 from itertools import zip_longest
+
+
+def test_simple_dict():
+    dict_str = 'beta1:0.9,beta2:0.999,epsilon:1e-8,lazy_update:true'
+    expected = {'beta1': 0.9, 'beta2': 0.999, 'epsilon': 1e-8, 'lazy_update': True}
+    parse = arguments.simple_dict()
+    assert parse(dict_str) == expected
+
 
 # note that while --prepared-data and --source/--target are mutually exclusive this is not the case at the CLI level
 @pytest.mark.parametrize("test_params, expected_params", [
@@ -31,6 +38,7 @@ from itertools import zip_longest
      '--output test_output',
      dict(source='test_src', target='test_tgt',
           source_factors=[],
+          source_factors_use_source_vocab=[],
           prepared_data='prep_data',
           validation_source='test_validation_src', validation_target='test_validation_tgt',
           validation_source_factors=[],
@@ -46,6 +54,7 @@ from itertools import zip_longest
      '-o test_output',
      dict(source='test_src', target='test_tgt',
           source_factors=[],
+          source_factors_use_source_vocab=[],
           prepared_data='prep_data',
           validation_source='test_validation_src', validation_target='test_validation_tgt',
           validation_source_factors=[],
@@ -60,16 +69,18 @@ def test_io_args(test_params, expected_params):
 
 
 @pytest.mark.parametrize("test_params, expected_params", [
-    ('', dict(quiet=False, loglevel='INFO')),
+    ('', dict(quiet=False,
+              loglevel='INFO',
+              no_logfile=False)),
 ])
 def test_logging_args(test_params, expected_params):
     _test_args(test_params, expected_params, arguments.add_logging_args)
 
 
 @pytest.mark.parametrize("test_params, expected_params", [
-    ('', dict(device_ids=[-1], use_cpu=False, disable_device_locking=False, lock_dir='/tmp')),
+    ('', dict(device_ids=[-1], use_cpu=False, omp_num_threads=None, disable_device_locking=False, lock_dir='/tmp')),
     ('--device-ids 1 2 3 --use-cpu --disable-device-locking --lock-dir test_dir',
-     dict(device_ids=[1, 2, 3], use_cpu=True, disable_device_locking=True, lock_dir='test_dir'))
+     dict(device_ids=[1, 2, 3], use_cpu=True, omp_num_threads=None, disable_device_locking=True, lock_dir='test_dir'))
 ])
 def test_device_args(test_params, expected_params):
     _test_args(test_params, expected_params, arguments.add_device_args)
@@ -81,12 +92,12 @@ def test_device_args(test_params, expected_params):
               num_layers=(6, 6),
               num_embed=(None, None),
               source_factors_num_embed=[],
-              source_factors_combine=C.SOURCE_FACTORS_COMBINE_CONCAT,
-              weight_tying=False,
-              weight_tying_type="trg_softmax",
+              source_factors_combine=[C.SOURCE_FACTORS_COMBINE_CONCAT],
+              source_factors_share_embedding=[False],
+              weight_tying_type="src_trg_softmax",
               transformer_attention_heads=(8, 8),
               transformer_feed_forward_num_hidden=(2048, 2048),
-              transformer_activation_type=C.RELU,
+              transformer_activation_type=(C.RELU, C.RELU),
               transformer_model_size=(512, 512),
               transformer_positional_embedding_type="fixed",
               transformer_preprocess=('n', 'n'),
@@ -95,7 +106,8 @@ def test_device_args(test_params, expected_params):
               encoder=C.TRANSFORMER_TYPE,
               decoder=C.TRANSFORMER_TYPE,
               dtype='float32',
-              amp=False))
+              amp=False,
+              amp_scale_interval=2000))
 ])
 def test_model_parameters(test_params, expected_params):
     _test_args(test_params, expected_params, arguments.add_model_parameters)
@@ -151,9 +163,9 @@ def test_inference_args(test_params, expected_params):
               checkpoint_improvement_threshold=0.,
               max_checkpoints=None,
               embed_dropout=(.0, .0),
-              transformer_dropout_attention=0.1,
-              transformer_dropout_act=0.1,
-              transformer_dropout_prepost=0.1,
+              transformer_dropout_attention=(0.1, 0.1),
+              transformer_dropout_act=(0.1, 0.1),
+              transformer_dropout_prepost=(0.1, 0.1),
               optimizer='adam',
               optimizer_params=None,
               horovod=False,
@@ -232,6 +244,7 @@ def test_tutorial_averaging_args(test_params, expected_params, expected_params_p
           source_vocab=None,
           target_vocab=None,
           source_factors=[],
+          source_factors_use_source_vocab=[],
           source_factor_vocabs=[],
           shared_vocab=False,
           num_words=(0, 0),
@@ -242,9 +255,12 @@ def test_tutorial_averaging_args(test_params, expected_params, expected_params_p
           no_bucket_scaling=False,
           max_seq_len=(99, 99),
           min_num_shards=1,
-          num_samples_per_shard=1000000,
+          num_samples_per_shard=10000000,
           seed=13,
-          output='train_data'
+          output='train_data',
+          quiet=False,
+          loglevel='INFO',
+          no_logfile=False
           ))
 ])
 def test_tutorial_prepare_data_cli_args(test_params, expected_params):
@@ -257,6 +273,7 @@ def test_tutorial_prepare_data_cli_args(test_params, expected_params):
           source_vocab=None,
           target_vocab=None,
           source_factors=[],
+          source_factors_use_source_vocab=[],
           source_factor_vocabs=[],
           shared_vocab=False,
           num_words=(0, 0),
@@ -267,9 +284,12 @@ def test_tutorial_prepare_data_cli_args(test_params, expected_params):
           no_bucket_scaling=False,
           max_seq_len=(99, 99),
           min_num_shards=1,
-          num_samples_per_shard=1000000,
+          num_samples_per_shard=10000000,
           seed=13,
-          output='prepared_data'
+          output='prepared_data',
+          quiet=False,
+          loglevel='INFO',
+          no_logfile=False
           ))
 ])
 def test_prepare_data_cli_args(test_params, expected_params):